requirements.txt
tkinter
matplotlib
pillow
mysql